Update .gitattributes.txt
For now I want Github's Linguist to count .h files as .swift files. This is hacky, but that way it will correctly display "Swift" next to the repo. The thing is that the public source files are indeed Objective-C, but that is because all Swift files are wrapped up in a xcframework.